feat:更新最新資料

+++ app/Http/Controllers/Api/V1/VehicleController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreVehicleRequest;
 use App\Http\Resources\VehicleResource;
 use App\Models\Vehicle;
+use Illuminate\Http\Response;
 
 /**
  * @group Vehicles
@@ -31,13 +32,25 @@ class VehicleController extends Controller
 
     public function update(StoreVehicleRequest $request, Vehicle $vehicle)
     {
+        if ($vehicle->hasActiveParkings()) {
+            return response()->json([
+                'errors' => ['general' => ['Can\'t update vehicle with active parkings. Stop active parking.']],
+            ], Response::HTTP_UNPROCESSABLE_ENTITY);
+        }
+
         $vehicle->update($request->validated());
 
-        return VehicleResource::make($vehicle);
+        return response()->json(VehicleResource::make($vehicle), Response::HTTP_ACCEPTED);
     }
 
     public function destroy(Vehicle $vehicle)
     {
+        if ($vehicle->hasActiveParkings()) {
+            return response()->json([
+                'errors' => ['general' => ['Can\'t remove vehicle with active parkings. Stop active parking.']],
+            ], Response::HTTP_UNPROCESSABLE_ENTITY);
+        }
+
         $vehicle->delete();
 
         return response()->noContent();
